Done with the about us and structurized the home-page
